[#12] feature : 주식 매매 기능 구현

+++ src/main/java/zb/accountMangement/account/dto/TransactionDto.java
@@ -14,17 +14,17 @@ import java.time.LocalDateTime;
 @Builder
 public class TransactionDto {
 
-  @NotNull
-  private Long accountId;
+    @NotNull
+    private Long accountId;
 
-  @NotBlank
-  private String name;
+    @NotBlank
+    private String name;
 
-  @NotBlank
-  private Long amount;
+    @NotBlank
+    private Double amount;
 
-  private String memo;
+    private String memo;
 
-  @CreatedDate
-  private LocalDateTime transactedAt;
+    @CreatedDate
+    private LocalDateTime transactedAt;
 }
